Error Could not create an instance of type Shesha.OPT.Dto.ISendPinInput occurs when sending an OTP #2310
