+++ shesha-core/src/Shesha.Framework/Otp/Dto/ResendPinInput.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace Shesha.Otp.Dto
 {
-    public class ResendPinInput
+    public class ResendPinInput: IResendPinInput
     {
         /// <summary>
         /// Unique runtime identifier of the operation
